Connected the weather chopper service.

+++ lib/data/providers/interfaces/app_weather_abstract.dart
@@ -2,6 +2,6 @@ import 'package:weather/models/forecast_weather/forecast_weather.dart';
 import 'package:weather/models/weather/weather.dart';
 
 abstract class AppWeatherAbstract {
-  Future<Weather> getCurrentWeather(int lat, int long);
-  Future<ForecastWeather> getForecastWeather(int lat, int long);
+  Future<Weather> getCurrentWeather(double lat, double long);
+  Future<ForecastWeather> getForecastWeather(double lat, double long);
 }
